% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-glm.R
\name{block.glm}
\alias{block.glm}
\title{block.glm}
\usage{
block.glm(f.lhs, f.rhs, data, order.by = NULL, strat.by = NULL,
  block.size = 20, B = 200, ...)
}
\arguments{
\item{f.lhs}{character vector, left hand side of a formula, the
model(s) to be fit will be defined by
`paste(f.lhs, f.rhs, sep=" ~ ")`}

\item{f.rhs}{character string, right hand side of a formula}

\item{data}{data.table containing the columns referred to in f.lhs
and f.rhs}

\item{order.by}{if not `NULL`, the name of a column in `data` on
which it should be sorted}

\item{strat.by}{if not `NULL`, the name of a column in `data` on
which it should be stratified before block sampling.  Eg, if
you are considering genomic data, you should stratify by
chromosome as there should be no spatial correlation between
chromosomes}

\item{block.size}{size of blocks of contiguous observations that
will be sampled for bootstrap estimation of variance of
parameter estimates}

\item{B}{number of bootstrap estimates}

\item{...}{other arguments passed to `glm()` (eg
`family="binomial"`)}
}
\value{
data.table giving the estimated effect ("beta") of each
    item in f.rhs on each item in f.lhs, together with block
    bootstrap estimates of confidence interval (beta.025,
    beta.975) and standard error (se.beta) and the number of
    bootstraps on which those estimates are based.
}
\description{
Regression models for genomic data often assume there is
independence between neighbouring genomic elements when, in
reality, there is spatial dependence.  This function implements a
block bootstrap method for estimating correct variances of
paramter estimates.
}
\details{
Note that this function uses `mclapply` to parallelise the
bootstrapping.  Please set `mc.cores` to something sensible, eg
\code{options(mc.cores=10)}
if you have 10 cores.
}
\author{
Chris Wallace and Oliver Burren
}

