% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-glm.R
\name{.sim.data}
\alias{.sim.data}
\title{internal function to simulate data for examples}
\usage{
.sim.data(n = 500, m = 10, beta = 0.2)
}
\arguments{
\item{n}{number of independent observations}

\item{m}{group size - number of times each independent observation is repeated}

\item{beta}{Y1 ~ N( beta * X, 1)}
}
\value{
data.table with
Chr (always 1, possibly needed for bootstrap),
x (explanatory variable),
y1 (response variable related to x),
y0 (response variable unrelated to x)
name (unique name for each independent observation)
}
\description{
internal function to simulate data for examples
}
\author{
Chris Wallace
}
